feat(book list): add book list template
Added template which renders books for user. Page contains button in which filters are shown, search template, pagination. Also description is truncated in list to avoid over extending item view.
Additionally, links to genres and types are added for each item to filter out books by the filter. Las but not lease, when there is nothing to show (list of books is empty), template shows message Nothing to show