feat(media user): add media user views
It includes detail, list, delete, update, create, update views and with custom logic: in list it's filtering query result by search field and including additional context to templates. In detail view that's adding context such as delete url (delete template is used across all models so it's easier to give path as context) and media type. in update view that's get_form (required for SetPasswordForm to apply new changes for user) and success url which logouts user and redirects to login page.